+++ include/models.hpp
@@ -1,21 +1,22 @@
 #ifndef RS_MODELS_HPP
 #define RS_MODELS_HPP
 
-#include "model.hpp"
-#include "field.hpp"
+#include "model/model.hpp"
+#include "model/field.hpp"
 
 namespace rs::model {
+
+/* Models for Database */
 struct User : Model {
-    Field<int,cnstr::Unique> id;
+    Field<unsigned long,cnstr::Unique> id;
     Field<std::string, cnstr::Unique, cnstr::Length<1,10>, cnstr::Required> username;
     Field<std::string, cnstr::Required, cnstr::Length<6,255>> password;
     Field<std::string, cnstr::Unique,cnstr::Required, cnstr::NotEmpty, cnstr::Length<2,32>> email;
     Field<std::string, cnstr::Required, cnstr::Length<2,64>> firstname;
     Field<std::string, cnstr::Required, cnstr::Length<2,64>> lastname;
-    Field<int, cnstr::Required> born;
+    Field<unsigned long, cnstr::Required> born;
     Field<std::string> status;
 };
-
 }
 
 /* Required for REFLECTION!! */
@@ -30,8 +31,6 @@ REFL_AUTO(
   field(born),
   field(status)
 )
-
-
 
 /* Required to enable automatic reading from sql to model and from model to sql */
 template <> struct soci::type_conversion<rs::model::User> : rs::model::specialize_model<rs::model::User> {};
